+++ OptAdvisor_Server/src/main/java/utf8/citicup/dataServiceImpl/UserDataServiceImpl.java
@@ -43,12 +43,21 @@ public class UserDataServiceImpl implements UserDataService{
     @Override
     @CacheEvict(value = "user")
     public void delete(String username) {
-        userRepository.deleteById(username);
+        if(userRepository.findByUsername(username)!=null) {
+            userRepository.deleteById(username);
+        }
     }
 
     @Override
     @Cacheable(value = "user")
     public User findById(String username) {
         return userRepository.findByUsername(username);
+    }
+
+    @Override
+    public boolean updatePassword(String username, String password) {
+        if(userRepository.findByUsername(username)==null)return false;
+        userRepository.updatePassword(username,password);
+        return true;
     }
 }
